Split dat.gui modules
